split rules.rs into one file per entoty
